Move GroupShape to quant_utils.py
