+++ vllm/model_executor/layers/quantization/compressed_tensors/schemes/compressed_tensors_24.py
@@ -11,12 +11,13 @@ from compressed_tensors.quantization import (QuantizationArgs,
 from compressed_tensors.utils import combine_shards
 
 from vllm import _custom_ops as ops
-from vllm.compilation.fusion import GroupShape
 from vllm.model_executor.layers.fp8_quantization import QuantFP8
 from vllm.model_executor.layers.linear import (MergedColumnParallelLinear,
                                                QKVParallelLinear)
 from vllm.model_executor.layers.quantization.compressed_tensors.schemes import (
     CompressedTensorsScheme)
+from vllm.model_executor.layers.quantization.utils.quant_utils import (
+    GroupShape)
 from vllm.model_executor.layers.quantization.utils.w8a8_utils import (
     convert_to_channelwise, sparse_cutlass_supported)
 from vllm.model_executor.parameter import (BasevLLMParameter,
