+++ vllm/model_executor/layers/quantization/utils/w8a8_utils.py
@@ -7,9 +7,10 @@ import torch
 
 from vllm import _custom_ops as ops
 from vllm import envs
-from vllm.compilation.fusion import GroupShape
 from vllm.config import CompilationLevel, get_current_vllm_config
 from vllm.model_executor.layers.fp8_quantization import QuantFP8
+from vllm.model_executor.layers.quantization.utils.quant_utils import (
+    GroupShape)
 from vllm.platforms import current_platform
 
 # Input scaling factors are no longer optional in _scaled_mm starting
